Conditionally display the entire RepCap subsection
Include the "RepCap" subsection header in the conditional block. Instead
of displaying a message that there were not enough reads to show the
plot, now it will not show anything for "RepCap" there.

Tested that this Markdown-based switch works correctly. The trailing
newlines are required since this string injects straight into the
Markdown source before rendering.